chore: move config merge endpoint tests into separate test with build tags
