+++ comp/core/autodiscovery/impl/BUILD.bazel
@@ -67,6 +67,7 @@ dd_agent_go_test(
         "autoconfig_test.go",
         "common_test.go",
         "configmgr_discovery_test.go",
+        "configmgr_kube_endpoints_test.go",
         "configmgr_test.go",
         "multimap_test.go",
         "secrets_test.go",
